refactor: use a common macro to define math UDFs
