+++ datafusion/functions/Cargo.toml
@@ -60,6 +60,7 @@ datafusion-physical-expr = { workspace = true, default-features = true }
 hex = { version = "0.4", optional = true }
 itertools = { workspace = true }
 log = { workspace = true }
+paste = "1.0.14"
 regex = { version = "1.8", optional = true }
 
 [dev-dependencies]
