+++ datafusion/functions/src/macros.rs
@@ -146,3 +146,96 @@ macro_rules! downcast_arg {
         })?
     }};
 }
+
+macro_rules! make_math_unary_function {
+    ($UDF:ty, $GNAME:ident, $NAME:ident, $UNARY_FUNC:ident) => {
+        make_udf_function!(paste::paste! {$NAME::[<$UDF>]}, $GNAME, $NAME);
+
+        mod $NAME {
+            use arrow::array::{ArrayRef, Float32Array, Float64Array};
+            use arrow::datatypes::DataType;
+            use datafusion_common::{exec_err, DataFusionError, Result};
+            use datafusion_expr::ColumnarValue;
+            use datafusion_expr::{ScalarUDFImpl, Signature, Volatility};
+            use std::any::Any;
+            use std::sync::Arc;
+
+            paste::paste! {
+                #[derive(Debug)]
+                pub struct $UDF {
+                    signature: Signature,
+                }
+            }
+
+            impl $UDF {
+                pub fn new() -> Self {
+                    use DataType::*;
+                    Self {
+                        signature: Signature::uniform(
+                            1,
+                            vec![Float64, Float32],
+                            Volatility::Immutable,
+                        ),
+                    }
+                }
+            }
+
+            impl ScalarUDFImpl for $UDF {
+                fn as_any(&self) -> &dyn Any {
+                    self
+                }
+                fn name(&self) -> &str {
+                    stringify!($NAME)
+                }
+
+                fn signature(&self) -> &Signature {
+                    &self.signature
+                }
+
+                fn return_type(&self, arg_types: &[DataType]) -> Result<DataType> {
+                    let arg_type = &arg_types[0];
+
+                    match arg_type {
+                        DataType::Float64 => Ok(DataType::Float64),
+                        DataType::Float32 => Ok(DataType::Float32),
+
+                        // For other types (possible values null/int), use Float 64
+                        _ => Ok(DataType::Float64),
+                    }
+                }
+
+                fn invoke(&self, args: &[ColumnarValue]) -> Result<ColumnarValue> {
+                    let args = ColumnarValue::values_to_arrays(args)?;
+
+                    let arr: ArrayRef = match args[0].data_type() {
+                        DataType::Float64 => {
+                            Arc::new(make_function_scalar_inputs_return_type!(
+                                &args[0],
+                                self.name(),
+                                Float64Array,
+                                Float64Array,
+                                { f64::$UNARY_FUNC }
+                            ))
+                        }
+                        DataType::Float32 => {
+                            Arc::new(make_function_scalar_inputs_return_type!(
+                                &args[0],
+                                self.name(),
+                                Float32Array,
+                                Float32Array,
+                                { f32::$UNARY_FUNC }
+                            ))
+                        }
+                        other => {
+                            return exec_err!(
+                                "Unsupported data type {other:?} for function {}",
+                                self.name()
+                            )
+                        }
+                    };
+                    Ok(ColumnarValue::Array(arr))
+                }
+            }
+        }
+    };
+}
